# perf: optimize check invalid field-vchar