--- a/benchmarks/core/is-valid-header-char.mjs
+++ b/benchmarks/core/is-valid-header-char.mjs
@@ -1,0 +1,57 @@
+import { bench, group, run } from 'mitata'
+import { isValidHeaderChar } from '../../lib/core/util.js'
+
+const html = 'text/html'
+const json = 'application/json; charset=UTF-8'
+
+const headerCharRegex = /[^\t\x20-\x7e\x80-\xff]/
+
+/**
+ * @param {string} characters
+ */
+function charCodeAtApproach (characters) {
+  // Validate if characters is a valid field-vchar.
+  //  field-value    = *( field-content / obs-fold )
+  //  field-content  = field-vchar [ 1*( SP / HTAB ) field-vchar ]
+  //  field-vchar    = VCHAR / obs-text
+  for (let i = 0; i < characters.length; ++i) {
+    const code = characters.charCodeAt(i)
+    // not \x20-\x7e, \t and \x80-\xff
+    if ((code < 0x20 && code !== 0x09) || code === 0x7f || code > 0xff) {
+      return false
+    }
+  }
+  return true
+}
+
+group(`isValidHeaderChar# ${html}`, () => {
+  bench('regexp.test', () => {
+    return !headerCharRegex.test(html)
+  })
+  bench('regexp.exec', () => {
+    return headerCharRegex.exec(html) === null
+  })
+  bench('charCodeAt', () => {
+    return charCodeAtApproach(html) === true
+  })
+  bench('isValidHeaderChar', () => {
+    return isValidHeaderChar(html) === true
+  })
+})
+
+group(`isValidHeaderChar ${json}`, () => {
+  bench('regexp.test', () => {
+    return !headerCharRegex.test(json)
+  })
+  bench('regexp.exec', () => {
+    return headerCharRegex.exec(json) === null
+  })
+  bench('charCodeAt', () => {
+    return charCodeAtApproach(json)
+  })
+  bench('isValidHeaderChar', () => {
+    return isValidHeaderChar(json)
+  })
+})
+
+await run()