--- a/lib/core/request.js
+++ b/lib/core/request.js
@@ -9,17 +9,6 @@ const util = require('./util')
 const { channels } = require('./diagnostics.js')
 const { headerNameLowerCasedRecord } = require('./constants')
 
-// headerCharRegex have been lifted from
-// https://github.com/nodejs/node/blob/main/lib/_http_common.js
-
-/**
- * Matches if val contains an invalid field-vchar
- *  field-value    = *( field-content / obs-fold )
- *  field-content  = field-vchar [ 1*( SP / HTAB ) field-vchar ]
- *  field-vchar    = VCHAR / obs-text
- */
-const headerCharRegex = /[^\t\x20-\x7e\x80-\xff]/
-
 // Verifies that a given path is valid does not contain control chars \x00 to \x20
 const invalidPathRegex = /[^\u0021-\u00ff]/
 
@@ -166,13 +155,12 @@ class Request {
           if (!Array.isArray(header) || header.length !== 2) {
             throw new InvalidArgumentError('headers must be in key-value pair format')
           }
-          const [key, value] = header
-          processHeader(this, key, value)
+          processHeader(this, header[0], header[1])
         }
       } else {
         const keys = Object.keys(headers)
-        for (const key of keys) {
-          processHeader(this, key, headers[key])
+        for (let i = 0; i < keys.length; ++i) {
+          processHeader(this, keys[i], headers[keys[i]])
         }
       }
     } else if (headers != null) {
@@ -315,7 +303,7 @@ class Request {
   }
 }
 
-function processHeader (request, key, val, skipAppend = false) {
+function processHeader (request, key, val) {
   if (val && (typeof val === 'object' && !Array.isArray(val))) {
     throw new InvalidArgumentError(`invalid ${key} header`)
   } else if (val === undefined) {
@@ -335,7 +323,7 @@ function processHeader (request, key, val, skipAppend = false) {
     const arr = []
     for (let i = 0; i < val.length; i++) {
       if (typeof val[i] === 'string') {
-        if (headerCharRegex.exec(val[i]) !== null) {
+        if (!util.isValidHeaderChar(val[i])) {
           throw new InvalidArgumentError(`invalid ${key} header`)
         }
         arr.push(val[i])
@@ -349,7 +337,7 @@ function processHeader (request, key, val, skipAppend = false) {
     }
     val = arr
   } else if (typeof val === 'string') {
-    if (headerCharRegex.exec(val) !== null) {
+    if (!util.isValidHeaderChar(val)) {
       throw new InvalidArgumentError(`invalid ${key} header`)
     }
   } else if (val === null) {