--- a/lib/core/request.js
+++ b/lib/core/request.js
@@ -5,7 +5,19 @@ const {
   NotSupportedError
 } = require('./errors')
 const assert = require('node:assert')
-const util = require('./util')
+const {
+  isValidHTTPToken,
+  isValidHeaderChar,
+  isStream,
+  destroy,
+  isBuffer,
+  isFormDataLike,
+  isIterable,
+  isBlobLike,
+  buildURL,
+  validateHandler,
+  getServerName
+} = require('./util')
 const { channels } = require('./diagnostics.js')
 const { headerNameLowerCasedRecord } = require('./constants')
 
@@ -44,7 +56,7 @@ class Request {
 
     if (typeof method !== 'string') {
       throw new InvalidArgumentError('method must be a string')
-    } else if (!util.isValidHTTPToken(method)) {
+    } else if (!isValidHTTPToken(method)) {
       throw new InvalidArgumentError('invalid request method')
     }
 
@@ -80,13 +92,13 @@ class Request {
 
     if (body == null) {
       this.body = null
-    } else if (util.isStream(body)) {
+    } else if (isStream(body)) {
       this.body = body
 
       const rState = this.body._readableState
       if (!rState || !rState.autoDestroy) {
         this.endHandler = function autoDestroy () {
-          util.destroy(this)
+          destroy(this)
         }
         this.body.on('end', this.endHandler)
       }
@@ -99,7 +111,7 @@ class Request {
         }
       }
       this.body.on('error', this.errorHandler)
-    } else if (util.isBuffer(body)) {
+    } else if (isBuffer(body)) {
       this.body = body.byteLength ? body : null
     } else if (ArrayBuffer.isView(body)) {
       this.body = body.buffer.byteLength ? Buffer.from(body.buffer, body.byteOffset, body.byteLength) : null
@@ -107,7 +119,7 @@ class Request {
       this.body = body.byteLength ? Buffer.from(body) : null
     } else if (typeof body === 'string') {
       this.body = body.length ? Buffer.from(body) : null
-    } else if (util.isFormDataLike(body) || util.isIterable(body) || util.isBlobLike(body)) {
+    } else if (isFormDataLike(body) || isIterable(body) || isBlobLike(body)) {
       this.body = body
     } else {
       throw new InvalidArgumentError('body must be a string, a Buffer, a Readable stream, an iterable, or an async iterable')
@@ -119,7 +131,7 @@ class Request {
 
     this.upgrade = upgrade || null
 
-    this.path = query ? util.buildURL(path, query) : path
+    this.path = query ? buildURL(path, query) : path
 
     this.origin = origin
 
@@ -167,9 +179,9 @@ class Request {
       throw new InvalidArgumentError('headers must be an object or an array')
     }
 
-    util.validateHandler(handler, method, upgrade)
+    validateHandler(handler, method, upgrade)
 
-    this.servername = util.getServerName(this.host)
+    this.servername = getServerName(this.host)
 
     this[kHandler] = handler
 
@@ -314,7 +326,7 @@ function processHeader (request, key, val) {
 
   if (headerName === undefined) {
     headerName = key.toLowerCase()
-    if (headerNameLowerCasedRecord[headerName] === undefined && !util.isValidHTTPToken(headerName)) {
+    if (headerNameLowerCasedRecord[headerName] === undefined && !isValidHTTPToken(headerName)) {
       throw new InvalidArgumentError('invalid header key')
     }
   }
@@ -323,7 +335,7 @@ function processHeader (request, key, val) {
     const arr = []
     for (let i = 0; i < val.length; i++) {
       if (typeof val[i] === 'string') {
-        if (!util.isValidHeaderChar(val[i])) {
+        if (!isValidHeaderChar(val[i])) {
           throw new InvalidArgumentError(`invalid ${key} header`)
         }
         arr.push(val[i])
@@ -337,7 +349,7 @@ function processHeader (request, key, val) {
     }
     val = arr
   } else if (typeof val === 'string') {
-    if (!util.isValidHeaderChar(val)) {
+    if (!isValidHeaderChar(val)) {
       throw new InvalidArgumentError(`invalid ${key} header`)
     }
   } else if (val === null) {