--- a/lib/core/util.js
+++ b/lib/core/util.js
@@ -495,6 +495,24 @@ function isValidHTTPToken (characters) {
   return true
 }
 
+/**
+ * @param {string} characters
+ */
+function isValidHeaderChar (characters) {
+  // Validate if val is a valid field-vchar.
+  //  field-value    = *( field-content / obs-fold )
+  //  field-content  = field-vchar [ 1*( SP / HTAB ) field-vchar ]
+  //  field-vchar    = VCHAR / obs-text
+  for (let i = 0; i < characters.length; ++i) {
+    const code = characters.charCodeAt(i)
+    // not \x20-\x7e, \t and \x80-\xff
+    if ((code < 0x20 && code !== 0x09) || code === 0x7f || code > 0xff) {
+      return false
+    }
+  }
+  return true
+}
+
 // Parsed accordingly to RFC 9110
 // https://www.rfc-editor.org/rfc/rfc9110#field.content-range
 function parseRangeHeader (range) {
@@ -516,7 +534,6 @@ kEnumerableProperty.enumerable = true
 module.exports = {
   kEnumerableProperty,
   nop,
-
   isDisturbed,
   isErrored,
   isReadable,
@@ -546,6 +563,7 @@ module.exports = {
   buildURL,
   addAbortListener,
   isValidHTTPToken,
+  isValidHeaderChar,
   isTokenCharCode,
   parseRangeHeader,
   nodeMajor,