--- a/lib/core/util.js
+++ b/lib/core/util.js
@@ -495,22 +495,22 @@ function isValidHTTPToken (characters) {
   return true
 }
 
+// headerCharRegex have been lifted from
+// https://github.com/nodejs/node/blob/main/lib/_http_common.js
+
+/**
+ * Matches if val contains an invalid field-vchar
+ *  field-value    = *( field-content / obs-fold )
+ *  field-content  = field-vchar [ 1*( SP / HTAB ) field-vchar ]
+ *  field-vchar    = VCHAR / obs-text
+ */
+const headerCharRegex = /[^\t\x20-\x7e\x80-\xff]/
+
 /**
  * @param {string} characters
  */
 function isValidHeaderChar (characters) {
-  // Validate if val is a valid field-vchar.
-  //  field-value    = *( field-content / obs-fold )
-  //  field-content  = field-vchar [ 1*( SP / HTAB ) field-vchar ]
-  //  field-vchar    = VCHAR / obs-text
-  for (let i = 0; i < characters.length; ++i) {
-    const code = characters.charCodeAt(i)
-    // not \x20-\x7e, \t and \x80-\xff
-    if ((code < 0x20 && code !== 0x09) || code === 0x7f || code > 0xff) {
-      return false
-    }
-  }
-  return true
+  return !headerCharRegex.test(characters)
 }
 
 // Parsed accordingly to RFC 9110