```
benchmark              time (avg)             (min … max)       p75       p99      p999
--------------------------------------------------------- -----------------------------
• isValidHeaderChar# text/html
--------------------------------------------------------- -----------------------------
regexp.test            34 ns/iter     (28.27 ns … 303 ns)  32.71 ns  72.85 ns    133 ns
regexp.exec         40.45 ns/iter     (34.18 ns … 218 ns)  41.16 ns  83.69 ns    135 ns
charCodeAt          30.31 ns/iter     (26.07 ns … 217 ns)   29.1 ns  61.91 ns    123 ns
isValidHeaderChar   35.55 ns/iter      (29.3 ns … 236 ns)  35.45 ns  86.28 ns    126 ns

summary for isValidHeaderChar# text/html
  charCodeAt
   1.12x faster than regexp.test
   1.17x faster than isValidHeaderChar
   1.33x faster than regexp.exec

• isValidHeaderChar application/json; charset=UTF-8
--------------------------------------------------------- -----------------------------
regexp.test         65.64 ns/iter     (56.88 ns … 426 ns)  63.33 ns    152 ns    237 ns
regexp.exec         73.79 ns/iter     (63.48 ns … 283 ns)  72.61 ns    153 ns    208 ns
charCodeAt            108 ns/iter     (95.21 ns … 416 ns)    102 ns    238 ns    353 ns
isValidHeaderChar   66.89 ns/iter     (57.03 ns … 254 ns)  63.48 ns    156 ns    207 ns

summary for isValidHeaderChar application/json; charset=UTF-8
  regexp.test
   1.02x faster than isValidHeaderChar
   1.12x faster than regexp.exec
   1.64x faster than charCodeAt
```